+++ Changelog.md
@@ -14,6 +14,7 @@
 - improved AI of baby enderdragon
     - can not take damage from own attacks any more (immune to own area effect clouds that do damage)
     - once tamed, no longer tries to shoot other players
+    - attack/follow range is now adjustable through the config
 
 ### Version 1.18.2 - 2.0.3